CAN BUS Araç Veri Paneli Web Sitesi/requirements.txt
absl-py==2.1.0
alembic==1.14.1
aplr==10.8.0
asgiref==3.9.1
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_cytoscape==1.0.2
debugpy==1.8.11
decorator==5.1.1
dill==0.3.9
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
efficientnet==1.1.1
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.12.23
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gevent==24.11.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h5py==3.12.1
idna==3.10
imageio==2.36.1
importlib_metadata==8.6.1
interpret==0.6.9
interpret-core==0.6.9
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.7.0
Keras-Applications==1.0.8
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multiprocess==0.70.17
namex==0.0.8
narwhals==1.27.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
optuna==4.2.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
retrying==1.3.4
rich==13.9.4
SALib==1.5.1
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
serial==0.0.97
setuptools==75.3.0
shap==0.46.0
sip==6.9.1
six==1.16.0
slicer==0.0.8
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.124
ultralytics-thop==2.0.10
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
